adding changes prior to abandoning batch queries
investigation of shell scripts appears to suggest query of sites one at
a time. MH reviews related function in dump.sh and agrees this to be
true. Furthermore, trial and error with api request with differing date
parameters appears to return 4224 rows, this appears to be the maximum
number of rows returned despite the MAXROWS object specifying 40000, as
in webtri.sh. Therefore I suggest adjusting MAXROWS to 4224 and querying
1 site at a time. hitting MAXROWS will require query pagination.